It's asynchronous now!
Mod help page icons to all use unicode.
Add favicon.ico and reference in index.html.
Add init.js to contain all startup and initialization Functions
Mod initialization to be in init.js instead of index.js
Mod script calls for index.js and init.js to be at the bottom of the index page, putting them in the head caused the page to reload on input submit.
Mod parameter of setDisplay to be 'data'.
Mod all of the formerly synchronous ajax requests to be asynchrnous and still work.
Add inconsolata to the font-family list, just after Ubuntu Mono.